feat: refactor transaction controller
