+++ src/main/java/com/rabobank/bankapplication/service/CategoryService.java
@@ -1,7 +1,6 @@
 package com.rabobank.bankapplication.service;
 
 import com.rabobank.bankapplication.model.Category;
-import com.rabobank.bankapplication.model.Transaction;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
