+++ src/main/java/com/rabobank/bankapplication/service/TransactionService.java
@@ -13,6 +13,7 @@ public interface TransactionService {
     List<Transaction> getTransactionsByAccountId(Long accountId);
 
     List<Transaction> getTransactionsByCategoryIdAndAccountId(Long accountId, Long categoryId);
+
     Transaction getTransactionById(Long id);
 
     Transaction createNewTransaction(Transaction transaction);
